python/requirements.txt
pyyaml>=6.0
pandas>=2.0
numpy>=1.24
scipy>=1.10
matplotlib>=3.7
seaborn>=0.12
stereopy>=1.6
